requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0
numpy==2.1.3
torch==2.4.1
torchvision==0.19.1
pytest==8.4.2
httpx==0.28.1
pytorch-ignite==0.5.3